pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["Q <phlacx@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
blacksheep = "^2.3.1"
uvicorn = "^0.34.0"
gnews = "^0.3.9"
googlenewsdecoder = "^0.1.6"
google-generativeai = "^0.8.3"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.23.0"}
newspaper3k = "^0.2.8"
python-dotenv = "^1.0.1"
pydantic = "^2.10.4"
python-daemon = "^3.1.2"
pydantic-settings = "^2.2.1"

[tool.poetry.scripts]
server = "scraper.main:server"
daemon = "scraper.main:run_as_daemon"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --cov=scraper --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["scraper"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["scraper"]
skip = [".gitignore", ".dockerignore"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
